Add useful MCP prompts and bounded resources
