--- .github/workflows/apply-lobehub-mcp-surfaces-pr.yml
@@ -1,46 +1,0 @@
-name: Apply LobeHub MCP Surfaces on Draft PR
-
-on:
-  pull_request:
-    paths:
-      - .github/scripts/apply_lobehub_mcp_surfaces.py
-      - .github/workflows/apply-lobehub-mcp-surfaces-pr.yml
-      - tests/test_mcp_discovery_surfaces.py
-
-permissions:
-  contents: write
-
-jobs:
-  apply:
-    if: github.repository == 'juyterman1000/entroly' && github.head_ref == 'feature/lobehub-score-remediation'
-    runs-on: ubuntu-latest
-    timeout-minutes: 25
-    steps:
-      - uses: actions/checkout@v6
-        with:
-          ref: feature/lobehub-score-remediation
-          fetch-depth: 0
-
-      - name: Apply anchored source patch
-        run: python .github/scripts/apply_lobehub_mcp_surfaces.py
-
-      - name: Validate discovery surfaces
-        run: |
-          python -m compileall -q entroly/server.py tests/test_mcp_discovery_surfaces.py
-          python -m pip install -q -e . 'mcp>=1.6.0' pytest
-          PYTHONPATH="$GITHUB_WORKSPACE" pytest tests/test_mcp_discovery_surfaces.py -q --tb=short
-
-      - name: Commit source and remove bootstrap helpers
-        shell: bash
-        run: |
-          git config user.name "juyterman1000"
-          git config user.email "208309368+juyterman1000@users.noreply.github.com"
-          git rm -f \
-            .github/scripts/apply_lobehub_mcp_surfaces.py \
-            .github/workflows/apply-lobehub-mcp-surfaces.yml \
-            .github/workflows/apply-lobehub-mcp-surfaces-v2.yml \
-            .github/workflows/apply-lobehub-mcp-surfaces-pr.yml \
-            .github/workflows/debug-lobehub-mcp-surfaces.yml
-          git add entroly/server.py tests/test_mcp_discovery_surfaces.py
-          git commit -m "Add useful MCP prompts and bounded resources"
-          git push origin HEAD:feature/lobehub-score-remediation
