--- .github/workflows/debug-lobehub-mcp-surfaces.yml
@@ -1,43 +1,0 @@
-name: Debug LobeHub MCP Surfaces
-
-on:
-  pull_request:
-    paths:
-      - .github/workflows/debug-lobehub-mcp-surfaces.yml
-
-permissions:
-  contents: read
-
-jobs:
-  debug:
-    if: github.head_ref == 'feature/lobehub-score-remediation'
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v6
-        with:
-          ref: feature/lobehub-score-remediation
-      - name: Apply source patch in workspace
-        run: python .github/scripts/apply_lobehub_mcp_surfaces.py
-      - name: Run focused validation with captured output
-        shell: bash
-        run: |
-          set +e
-          python -m compileall -q entroly/server.py tests/test_mcp_discovery_surfaces.py > "$RUNNER_TEMP/compile.log" 2>&1
-          compile_status=$?
-          python -m pip install -q 'mcp>=1.6.0' pytest > "$RUNNER_TEMP/install.log" 2>&1
-          install_status=$?
-          pytest tests/test_mcp_discovery_surfaces.py -vv --tb=long > "$RUNNER_TEMP/pytest.log" 2>&1
-          pytest_status=$?
-          printf 'compile=%s\ninstall=%s\npytest=%s\n' "$compile_status" "$install_status" "$pytest_status" > "$RUNNER_TEMP/status.txt"
-          exit "$pytest_status"
-      - name: Upload diagnostics
-        if: always()
-        uses: actions/upload-artifact@v4
-        with:
-          name: lobehub-mcp-surface-debug
-          path: |
-            ${{ runner.temp }}/compile.log
-            ${{ runner.temp }}/install.log
-            ${{ runner.temp }}/pytest.log
-            ${{ runner.temp }}/status.txt
-          if-no-files-found: error
